refactor(labs): Add index file to simplify theming imports
